Şehir Plaka Sorgusu
This project is developed using the Kotlin programming language with the primary goal of determining the city associated with a given license plate number entered by the user. The user can input the license plate number through the application interface and initiate the process by clicking the 'Find City' button. The project accomplishes this matching by utilizing a database that associates license plate numbers with cities. Additionally, it provides a modern and efficient solution through the use of the Kotlin programming language. The aim is to create a user-friendly interface and ensure simple usage for everyone to easily comprehend.